rewrite UploadHandler, correct Android upload handling

+++ android/src/main/java/de/wellenvogel/avnav/appapi/JavaScriptApi.java
@@ -291,7 +291,7 @@ public class JavaScriptApi {
     }
 
     @JavascriptInterface
-    public boolean interruptCopy(int id){
+    public boolean interruptCopy(long id){
         if (uploadData==null || ! uploadData.isReady(id)) return false;
         return uploadData.interruptCopy(false);
     }
